Port nonzero tests to new integration test style
