--- schemars/src/json_schema_impls/nonzero_signed.rs
@@ -1,37 +1,0 @@
-use crate::SchemaGenerator;
-use crate::_alloc_prelude::*;
-use crate::{JsonSchema, Schema};
-use alloc::borrow::Cow;
-use core::num::*;
-
-macro_rules! nonzero_unsigned_impl {
-    ($type:ty => $primitive:ty) => {
-        impl JsonSchema for $type {
-            always_inline!();
-
-            fn schema_name() -> Cow<'static, str> {
-                stringify!($type).into()
-            }
-
-            fn schema_id() -> Cow<'static, str> {
-                stringify!(std::num::$type).into()
-            }
-
-            fn json_schema(generator: &mut SchemaGenerator) -> Schema {
-                let mut schema = <$primitive>::json_schema(generator);
-                let object = schema.ensure_object();
-                object.insert("not".to_owned(), serde_json::json!({
-                    "const": 0
-                }));
-                schema
-            }
-        }
-    };
-}
-
-nonzero_unsigned_impl!(NonZeroI8 => i8);
-nonzero_unsigned_impl!(NonZeroI16 => i16);
-nonzero_unsigned_impl!(NonZeroI32 => i32);
-nonzero_unsigned_impl!(NonZeroI64 => i64);
-nonzero_unsigned_impl!(NonZeroI128 => i128);
-nonzero_unsigned_impl!(NonZeroIsize => isize);
