--- schemars/tests/expected/nonzero_ints.json
@@ -1,34 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "MyStruct",
-  "type": "object",
-  "properties": {
-    "unsigned": {
-      "type": "integer",
-      "format": "uint32",
-      "minimum": 0
-    },
-    "nonzero_unsigned": {
-      "type": "integer",
-      "format": "uint32",
-      "minimum": 1
-    },
-    "signed": {
-      "type": "integer",
-      "format": "int32"
-    },
-    "nonzero_signed": {
-      "type": "integer",
-      "format": "int32",
-      "not": {
-        "const": 0
-      }
-    }
-  },
-  "required": [
-    "unsigned",
-    "nonzero_unsigned",
-    "signed",
-    "nonzero_signed"
-  ]
-}
